Add files via upload
Added a new background and switched the main player from the balloon to a video game character. When the player reaches a certain coordinate, it switches backgrounds. Didn't finish and ran into issues with blitting the new background so code doesn't work at the moment but plan on polishing tomorrow.