Add files via upload
I used the majority of my time to fix my bug today and was able to sort of fix it through screen.fill to prevent the two backgrounds from overlapping, however my objects (tree and house) continue to glitch through the background. But now that I’ve mainly fixed this issue, I can move on to adding other characters and dialogue. I sort of began rendering and creating storyline + dialogues for the characters.