Add files via upload
Over the weekend I worked on my code a lot. I was able to add dialogue and include multiple backgrounds that the program moves in between when the user makes a certain choice. Additionally I tried to make a class with functions that prints out texts letter by letter. I watched many videos and websites to learn to implement that into my code. All that’s left is adding the characters.